登录添加返回用户名

+++ src/main/java/com/hospital_care/hospitalcare/entity/role_permission/vo/PermissionRoleParmVo.java
@@ -1,6 +1,7 @@
 package com.hospital_care.hospitalcare.entity.role_permission.vo;
 
 import com.hospital_care.hospitalcare.entity.permission.vo.TreeVo;
+import io.swagger.models.auth.In;
 import lombok.Data;
 
 import java.util.List;
@@ -8,5 +9,5 @@ import java.util.List;
 @Data
 public class PermissionRoleParmVo {
     private List<TreeVo> list;
-    private Long roleId;
+    private Integer roleId;
 }
